rearrange chain hierarchy
